Add CSV file support for matchings/actions.
Matchings/actions can now use values read from a
Comma Separated Value (CSV) file.  This makes it
easier to integrate E9Tool with other tools.

For example, it is possible to define a file
called data.csv as follows:

    0x434b67,111
    0x434b6c,222
    0x434b55,333
    ...

Then we can use the match the first column
against the instruction address, and the second
column can be passed as an argument to call
instrumentation.  For example:

    ./e9tool --match  'addr=data[0]'
             --action 'call f(data[1])@bin'
             ...

For example, the value 222 will be passed to
f() for the instruction at address 0x434b6c,
etc.

This feature is not well tested and should be
considered unstable.

Also fixes #6.