Implement register pass-by-pointer arguments.
E9Tool now supports passing registers values by
pointer, e.g.: &rax, &r15, ..., etc.

This also allows the instrumentation function to
modify the register contents if desired.

It is implemented by allocating a stack slot for
each register.  The slot is initialized to the
register value before the call, and the register
is set to the slot value after the call.