Implement operand src[i]/dst[i]/op[i] arguments.
Implement src[i]/dst[i]/op[i] call arguments.
E.g.:

  ./e9tool -M ...
           -A 'call f(src[0],src[1],dst[0])@bin'
           ...

The arguments will be the first/second source
operands, followed by the first destination
operand.

Operands are represented by pointers:
- Register operands: a pointer to memory that
  represents the register.  Similar to &rax, etc.
- Memory operands: a pointer to the memory.
- Immediate operands: a (const) pointer the
  immediate value.

For non-read-only operands, the pointer can be
written to, and this will change the CPU state.
This allows for state-modification.

If operands alias then so will the pointers.

This is a work-in-progress.
Implementing this required call instrumentation
to be redesigned.  The new implementation will
likely be unstable for a while.