Add libc support for call instrumentation.
Adds a parallel libc implementation in the form
of stdlib.c.  To use, #include the file:

    #include "stdlib.c"

This is a parallel implementation and not the
"real" libc, which cannot be safely invoked
from call instrumentation for many reasons:

    - SYSV vs clean ABI issues
    - transparency issues
      (e.g., changing errno, etc.)
    - deadlocks

The parallel implementation is designed to
address these issues, and provides a somewhat
normal programming environment for call
instrumentation (making it much easier to write
code).

Note the implementation is a WIP and likely
contains bugs.