Implement a conditional jump call kind.
Syntax:

    call[conditional.jump] func(...)@binary

Semantics:

    Inspects the return value (result) of func().
    If non-zero, then jump to result.
    Otherwise, execute the matching instruction
    (like before).

This fixes a limitation where it was very
difficult to implement control-flow-transfers
using call actions.

NOTE: This commit also swaps the logic of the
conditional call kind.  Previously, zero means do
not execute the matching instruction, and non-zero
means execute.  This has been swapped: so zero now
means execute and non-zero means do not execute.
This is to be consistent with conditional.jump.