Add -Ojump-delay optimization for E9Patch.
This optimization can significantly speed up the
patched binary.

The basic idea is as follows: suppose we have the
instruction sequence: I0, I1, I2, I3, I4, I5, ...
And assume that I0 is to be patched, then:
(1) I0 is replaced with a jump to a trampoline T0
(2) after T0 has finished, T0 jumps back to I1.

The basic idea of the -Ojump-delay optimization
is to avoid (2) if possible.  To do so, we search
the instructions I1,I2,..., to find the first
control-flow-transfer (CFT) instruction, e.g.,
jump/call/ret/etc. (incl. jumps to other
trampolines).  Next we copy all instructions up to
the CFT, and use the CFT to "return" from the
trampoline.

For example, say I5=callq is the first CFT
instruction.  Then we copy I1,...,I5 to the
trampoline, and use the callq instruction to
"return" from the trampoline.  Thus, we avoid the
original jump to I1 from (2).

This optimization can significantly reduce the
number of jumps used by the patched binary.  This
also increases locality and seems to be much
faster for some benchmarks.

The implementation is experimental so the
optimization is disabled by default.  To use, pass
the following options to E9Tool:

    --option -Ojump-delay=32