Add a simple basic-block analysis to E9Tool
This partly integrates the basic-block recovery
analysis from E9AFL into E9Tool.  The idea is
that the analysis is useful for other
applications beyond AFL instrumentation.

This is a WIP and not yet documented.  The
final form is not decided, and may change.

The analysis is designed to be a conservative
overapproximation for binaries generated by
"normal" compilers.  This means the analysis
may incorrectly flag some instructions as
basic-block entries, when they are really not.
For some applications this is tolerable.

The analysis does not attempt to solve the
undecidable general case.  The analysis is NOT
used for the rewriting itself.  Rather, the
binary rewriting is still control-flow agnostic
(E9Patch).