Allow trampoline templates in "patch" messages.
This change means that the "patch" message can
use ad hoc trampolines defined in the template
format, rather define a named trampoline in
advance.

It also now treats "trampoline" messages as
macros during macro expansion, in addition to
the existing "metadata".  This also means that
trampoline names must now start with a $.