Implement a new "trampoline prologue" optimization
The optimization works by relocating some
instructions *before* the patch location into a
"trampoline prologue".  The prologue essentially
serves as a "landing pad" for the -Opeephole
optimization.

The idea is similar to the -Oepilogue optimization
(formally -Ojump-elim).  Overall, this can
reduce the number of jumps the patched binary
uses.  However, -Oepilogue is more impactful.

The optimization is considered experimental and
is currently only enabled using E9Tool's -O3.

Other changes:
- Some e9patch options renamed.
- E9Tool now supports a --seed option.