Allow PT_GNU_RELRO/STACK use for the loader.
E9Patch uses the PT_NOTE trick in order to inject
the loader.  However, some binaries do not have a
PT_NOTE phdr.  This change allows PT_GNU_RELO or
PT_GNU_STACK to be used instead.

This must be enabled via the E9Ptch --phdr-loader
option.  It is disabled by default, since using
PT_GNU_* will weaken memory protections in the
rewritten binary, which may be a security risk.
For some applications, this may not be a concern.