Make trampolines composable.
Previously, there could only be ONE trampoline per
matching instruction.  E.g.:

    ./e9tool -M m1 -A print -M m2 -A 'call ...'

If an instruction I matches *both* m1 and m2, then
E9Tool will only choose the FIRST trampoline
(i.e., print), and call trampoline will not be
used.

This commit changes the semantics so that the
instruction I will be instrumented with *both* the
print AND call trampolines executed in order.

This makes it easy to compose instrumations, e.g.,
AFL instrumentation can be composed with print
instrumentation for debugging, etc.  Previously
it was highly non-trivial to compose trampolines.

This is a complex change and a WIP.  The current
commit breaks some features, like plugins, which
need to be redesigned and fixed later.