Refactor jump optimization to fix rare bug
Previously, jumps are optimized on-the-fly.  But
sometimes the optimized jump will target a
location that is not generated, because E9Patch
failed to relocate instructions to a concrete
address.  This resulted in a dangling jump+crash

Fixing the bug required a refactoring.  Now jumps
are optimized after than trampoline bytes are
generated, meaning that the existence of locs
are known when the optimization is applied.