Upgrde to a more modern version of Zydis.
This fixes some Zydis-related bugs in the old
version.

This also changes how the %cs/%es/%ss/%ds segment
registers are treated.  These registers are
ignored by x64, so are no longer decoded by
Zydis.  E.g.:

    movq %cs:-0x100(%rsp,%rsi,8), %rax

Will now be decoded the same as:

    movq -0x100(%rsp,%rsi,8), %rax

This appears to be the same treatment as XED.

Some of regtests have been updated to use %gs
instead, which is decoded.