Binary Rewriting *WITH* Control-Flow Recovery
This change makes it possible to rewrite binaries
using Control-Flow Recovery (CFR).

To enable, simply pass "-X" to E9Tool:

    $ ./e9tool -X ...

This will enable a heuristic-based CFR analysis
in E9Patch, and the binary will be rewritten
using this information.  Without the "-X", the
binary will be rewritten without CFR, which is
the classic E9Patch mode.

PROS (with -X):

- The rewritten binary will be much faster
- The patching coverage will be much higher
- The rewriting speed will be improved

CONS (with -X):

- If the heuristic-based CFR is inaccurate, the
  rewritten binary may be broken/buggy.
  Thus, the -X mode provides much weaker
  guarantees.

That said, the CFR analysis is designed to be
an overapproximation that should work for all
binaries compiled with standard compilers such
as gcc or clang.  The CFR does *NOT* assume
disassembly/symbols and can handle classic
problems like data-in-code.  It appears to work
for most binaries.

(The only incompatible binary found thus far is
 E9Patch itself, which contains code that
 delibrately attempts to confuse CFR analysis).

Whether the performance vs. accuracy trade-off
is acceptable depends on the application.  The
-X mode is generally much faster, and closes the
performance gap with inline-based binary
rewriting tools like RetroWrite.

Internally, CFR enables a new rewriting tactic
"T0" that batches together instructions that
are not jump targets into a single trampoline.
This significantly improves locality and
reduces jumps, improving performance.  This
also improves patching coverage, since
instruction batches are easier to patch, and
generally do not need to rely on instruction
punning.