Initial implementation of a "100% coverage" mode
This mode can be enabled using the -100 option:

    $ e9tool -100 ...

This mode attempts to get 100% instruction
coverage at the cost of performance.

To do so, the E9Patch backend implements a last
resort "B0" tactic which just replaces the
instruction with a single-byte illegal opcode.
When executed, the illegal opcode generates a
SIGILL which is then caught and control-flow
is then redirected to the trampoline.

This idea is not new, and since it uses signal
handlers, the performance overheads can be
significant.  However, for some applications,
it is desirable to have 100% coverage even if
the performance suffers.

This is not the "default" mode of E9Patch and
it is considered an extension.

It is still possible that 100% coverage may not
be achieved due to other reasons, such as if the
virtual address space is exhausted and there is
no more room for trampolines.

The current implemention is a WIP and needs work.