Add experimental E9Tool source file:line support
If the binary was compiled with debug information
(-g), we can use file:line information for both
matching and patching.  This initial commit adds
some basic (experimental) support.

Parsing the debug information is very difficult,
so we outsource this to libdw.so.  However, this
will be a "soft" dependency, i.e., E9Tool will
run fine if this library is not installed, but
then also file:line information will be missing.

This will only work if the binary was compiled
with (-g), otherwise the information will be
"undefined".